refactor: simplify @previewjs/app to bare minimum (#1219)
This will make it significantly easier to iterate on the app, as there
won't be a different codebase for Free / Pro versions, and no need to
extract common code anymore. Instead, they will be the same codebase
(which will now be closed-source).

The rest of Preview.js remains open source!